Create NoteTrainer class. Move show_devices to own class and file
